# scca-rest/src/main/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample.",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Sample",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "${scca.rest.context-path}/project" : {
      "get" : {
        "tags" : [ "Project MGT（项目管理）" ],
        "summary" : "List Project / 项目列表",
        "description" : "",
        "operationId" : "findAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespListProjectDto"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Project MGT（项目管理）" ],
        "summary" : "Create Project / 创建项目",
        "description" : "",
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ProjectDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Project MGT（项目管理）" ],
        "summary" : "Update Project / 更新项目",
        "description" : "",
        "operationId" : "updateProject",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ProjectDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Project MGT（项目管理）" ],
        "summary" : "Delete Project / 删除项目",
        "description" : "",
        "operationId" : "deleteProject",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/project/detail" : {
      "get" : {
        "tags" : [ "Project MGT（项目管理）" ],
        "summary" : "Project Detail / 项目详情",
        "description" : "",
        "operationId" : "findProjectDetail",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespProjectDto"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/project/env" : {
      "post" : {
        "tags" : [ "Project MGT（项目管理）" ],
        "summary" : "Project Add Env / 项目增加环境",
        "description" : "",
        "operationId" : "addProjectEnv",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Project MGT（项目管理）" ],
        "summary" : "Project Remove Env / 项目移除环境",
        "description" : "",
        "operationId" : "deleteProjectEnv",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/project/label" : {
      "post" : {
        "tags" : [ "Project MGT（项目管理）" ],
        "summary" : "Project Add Label / 项目增加配置版本",
        "description" : "",
        "operationId" : "addProjectLabel",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "labelName",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespLabelDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Project MGT（项目管理）" ],
        "summary" : "Project Delete Label / 项目删除版本",
        "description" : "",
        "operationId" : "deleteProjectLabel",
        "parameters" : [ {
          "name" : "labelId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/property/configServer" : {
      "get" : {
        "tags" : [ "Property MGT（配置详细管理）" ],
        "summary" : "config server properties / 获取配置中心获取的配置信息",
        "description" : "",
        "operationId" : "propertiesFromConfigServer",
        "parameters" : [ {
          "name" : "project",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "profile",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "label",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespListPropertySource"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/property/decrypt" : {
      "post" : {
        "tags" : [ "Property MGT（配置详细管理）" ],
        "summary" : "Decrypt property / 解密配置",
        "description" : "",
        "operationId" : "decrypt",
        "parameters" : [ {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/property/encrypt" : {
      "post" : {
        "tags" : [ "Property MGT（配置详细管理）" ],
        "summary" : "Encrypt property / 加密配置",
        "description" : "",
        "operationId" : "encrypt",
        "parameters" : [ {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/property/persistent" : {
      "get" : {
        "tags" : [ "Property MGT（配置详细管理）" ],
        "summary" : "persistent properties / 获取持久化的配置信息",
        "description" : "",
        "operationId" : "propertiesFromPersistent",
        "parameters" : [ {
          "name" : "project",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "profile",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "label",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespProperties"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Property MGT（配置详细管理）" ],
        "summary" : "Save properties / 保存持久化的配置信息",
        "description" : "",
        "operationId" : "saveProperties",
        "parameters" : [ {
          "name" : "project",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "profile",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "label",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "EnvDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "registryAddress" : {
          "type" : "string"
        },
        "configServerName" : {
          "type" : "string"
        },
        "contextPath" : {
          "type" : "string"
        }
      }
    },
...
    "LabelDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
...
    "ProjectDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "envs" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/EnvDto"
          }
        },
        "labels" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LabelDto"
          }
        }
      }
    },
    "PropertySource" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string",
          "readOnly" : true
        },
        "source" : {
          "type" : "object",
          "readOnly" : true,
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
...
    "WebRespLabelDto" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "$ref" : "#/definitions/LabelDto"
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
    "WebRespListProjectDto" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProjectDto"
          }
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
    "WebRespListPropertySource" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PropertySource"
          }
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
    "WebRespProjectDto" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "$ref" : "#/definitions/ProjectDto"
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
    "WebRespProperties" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
    "WebRespString" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
  }
}